Refactor: Update Koin ViewModel injection from `inject()` to `activityViewModel()` or `viewModel()`
This commit updates the Koin dependency injection for ViewModels across multiple fragments and MainActivity. It replaces the `inject()` delegate with `activityViewModel()` for shared ViewModels This change aligns with the recommended Koin practices for Android ViewModel injection.

Additionally, in `ShoeStoreApp.kt`, the `MainViewModel` registration is changed from `singleOf(::MainViewModel)` to `viewModelOf(::MainViewModel)` to reflect its lifecycle as a ViewModel.